app-misc/anki: unkeyword 25.02.6 for ~x86
The ~x86 keyword is a remnant of older versions, back when the program's
GUI presumably compiled on that arch. This hasn't been the case since
version 2.1.17, which introduced the dependency on net-libs/nodejs.

Since the GUI is arguably the most important part of Anki and I haven't
heard of anyone who runs Anki on (virtualized) x86 hardware nowadays,
I'm dropping the keyword.

Signed-off-by: Lucio Sauer <watermanpaint@posteo.net>